% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_utils.R
\name{build_request}
\alias{build_request}
\title{Build a httr2 request with optional auth}
\usage{
build_request(x, path, query = list(), headers = list())
}
\description{
Build a httr2 request with optional auth
}
\keyword{internal}
