% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{col_or_null}
\alias{col_or_null}
\title{Safe column extract that doesn't warn on missing tibble columns}
\usage{
col_or_null(df, col)
}
\description{
Safe column extract that doesn't warn on missing tibble columns
}
\keyword{internal}
