% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{hydro_service}
\alias{hydro_service}
\title{Create a hydro_service object}
\usage{
hydro_service(provider_id)
}
\arguments{
\item{provider_id}{ID as listed by \code{\link[=hydro_services]{hydro_services()}}}
}
\description{
Create a hydro_service object
}
