% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_registry.R
\name{hydro_services}
\alias{hydro_services}
\title{List available providers}
\usage{
hydro_services()
}
\value{
A tibble with columns:
provider_id, provider_name, country, base_url,
license, license_link, access_class, reuse_class, is_open_data
}
\description{
List available providers
}
