% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{normalize_utf8}
\alias{normalize_utf8}
\title{Normalize likely-misencoded UTF-8 text (fix common "Å" mojibake)}
\usage{
normalize_utf8(x)
}
\description{
Normalize likely-misencoded UTF-8 text (fix common "Å" mojibake)
}
\keyword{internal}
