% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/http_utils.R
\name{perform_request}
\alias{perform_request}
\title{Perform request with retry and exponential backoff}
\usage{
perform_request(req, max_tries = 5)
}
\description{
Perform request with retry and exponential backoff
}
\keyword{internal}
