% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{stations}
\alias{stations}
\title{List stations for a provider}
\usage{
stations(x, ...)
}
\arguments{
\item{x}{A \code{hydro_service} object created by \code{\link[=hydro_service]{hydro_service()}}.}

\item{...}{Passed to provider-specific methods.}
}
\value{
A tibble with station metadata.
}
\description{
List stations for a provider
}
