% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{timeseries_parameters}
\alias{timeseries_parameters}
\title{List supported parameters/units for a provider}
\usage{
timeseries_parameters(x, ...)
}
\arguments{
\item{x}{A \code{hydro_service} object.}

\item{...}{Reserved for future use.}
}
\value{
A tibble with columns: parameter, code, unit.
}
\description{
List supported parameters/units for a provider
}
