% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{to_ascii}
\alias{to_ascii}
\title{Transliterate to ASCII (remove diacritics); fallback to original on failure}
\usage{
to_ascii(x)
}
\description{
Transliterate to ASCII (remove diacritics); fallback to original on failure
}
\keyword{internal}
