% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_registry.R
\docType{data}
\name{.hydro_registry}
\alias{.hydro_registry}
\title{Internal registry for available providers}
\format{
An object of class \code{environment} of length 31.
}
\usage{
.hydro_registry
}
\description{
Internal registry for available providers
}
\keyword{internal}
