% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-pl_imgw_meta.R
\name{pl_imgw_meta}
\alias{pl_imgw_meta}
\title{IMGW station metadata (packaged)}
\format{
A tibble with one row per station.
}
\source{
Provided directly by IMGW to GRDC.
}
\description{
Enriched station metadata provided by IMGW (see \code{attr(., "source_stamp")} for the source timestamp).
Columns (tibble): station_id, station_name, river, lat_md, lon_md, area_md,
altitude_md, vertical_datum_md.
}
\keyword{datasets}
\keyword{internal}
